Add the login endpoint with the custom JWT support
For an authorization service I decided to dive in the JSON web tokens and
then implement them in the service. Now I have a refresh token that
stores in the database and an access token that doesn't store enywhere.

As a password hasher I chose argon2 (https://en.wikipedia.org/wiki/Argon2).

Whithin the commit I:
- Created an initial login logic
- Created a JWT_helper that incapsulates JWT logic
- Refactored some moments